esm - convert nls.js to nls.ts
