+++ src/tsconfig.json
@@ -38,7 +38,6 @@
 		"./main.js",
 		"./server-main.js",
 		"./server-cli.js",
-		"./vs/base/node/nls.js",
 		"./vs/base/parts/sandbox/electron-sandbox/preload-aux.js",
 		"./vs/base/parts/sandbox/electron-sandbox/preload.js",
 		"./vs/code/electron-sandbox/processExplorer/processExplorer.js",
