--- src/vs/base/node/nls.d.ts
@@ -1,38 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import type { INLSConfiguration } from 'vs/nls';
-
-export interface IResolveNLSConfigurationContext {
-
-	/**
-	 * Location where `nls.messages.json` and `nls.keys.json` are stored.
-	 */
-	readonly nlsMetadataPath: string;
-
-	/**
-	 * Path to the user data directory. Used as a cache for
-	 * language packs converted to the format we need.
-	 */
-	readonly userDataPath: string;
-
-	/**
-	 * Commit of the running application. Can be `undefined`
-	 * when not built.
-	 */
-	readonly commit: string | undefined;
-
-	/**
-	 * Locale as defined in `argv.json` or `app.getLocale()`.
-	 */
-	readonly userLocale: string;
-
-	/**
-	 * Locale as defined by the OS (e.g. `app.getPreferredSystemLanguages()`).
-	 */
-	readonly osLocale: string;
-}
-
-export function resolveNLSConfiguration(context: IResolveNLSConfigurationContext): Promise<INLSConfiguration>;
